LayoutTests/fast/sandbox/ios/focus-ring-color-expected.txt

Test that the focus ring color is as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.focusRingColor() is 'rgb(0, 122, 255)'
